+++ src/runcrew/providers/fit/__init__.py
@@ -1,0 +1,12 @@
+from runcrew.providers.fit.download import FitDownloadError, FitFileArtifact, FitFileStore
+from runcrew.providers.fit.parser import FitParseError, FitParseResult, parse_fit_activity
+
+__all__ = [
+    "FitDownloadError",
+    "FitFileArtifact",
+    "FitFileStore",
+    "FitParseError",
+    "FitParseResult",
+    "parse_fit_activity",
+]
+